bash_completion.d dir found in /etc not /usr/local
I found this page useful for setting up npm command completion. Thank you npm 
documentation team.

This suggested changed is based on my system which runs Ubuntu where
the completion directory path for bash is found in 
/etc/bash_completion.d and not /usr/local/etc/bash_completion.d

I dug around docs for this in The Bash Manual 
https://www.gnu.org/software/bash/manual/bashref.html#Programmable-Completion 
and on some an old (sadly retired) Debian Admin Site
https://debian-administration.org/article/317/An_introduction_to_bash_completion_part_2

which seemed supportive of the dir being in /etc but I do not have access 
to an OS X machine where I guess bash_completion.d could be in /usr/local/etc 
so I left it in.