Use native fs.writeFile in lib/install/save.js
To write a file, write-file-atomic uses a temp file and replaces the original one by renaming the temp file to its original's name. 
In some cases this produces an EBUSY error, if the original file cannot be removed/replaced.
By using fs.writeFile this issue can be avoided, since it writes the given content directly into the original file and hence does not try to replace it.

Correct me if I'm wrong but I don't think there is a specific downside in using the native writeFile here.