cli,outdated: default homepage to an empty string
This patch default package's homepage to an empty string in case the
homepage isn't specificied in the package manifest.

Before this patch, if at least one package didn't set an associated
homepage, running npm outdated --long would break.

After this patch, it won't display anything for this specific package
Homepage row and it won't break for the other packages.