ls: show installed but unmet peer deps
Any completely missing peer deps are filled in in the package tree, but
with only a name and the version with which it is required, and some
additional, derived metadata. However, if the dep does exist, but simply
is the wrong version, it is still marked as a missing peer dep.

This patch makes it so that `npm ls` does not hide *all* "missing" peer
deps, but just the ones that are not installed in some way already.

See https://npm.community/t/4770