Use Installer when packaging git dependencies
This fixes a couple of config issues we've found when installing
packages in `dry-run` mode to only get the lockfile changes.

When setting `dry-run` or `ignore-scripts`, the setting isn't
passed through when installing git dependencies meaning the `prepare`
script is always run.

We've also found an issue with proxying existing npm config to the cli
when a large `cafile` has been set, as this is expanded into the `ca`
config key with the full string contents of the ca file. If the ca file
is huge this will blow up on linux with an E2BIG exception (Argument
list too long). Increasing the limit requires kernel recompilation so
isn't really an option.

The current implementation requires `prepublish` to be fully deprecated,
is this something you are planning? Otherwise I'm happy to find another
way of passing this config through the installer.