test: check cache for root-owned files
This requires that all tests use `common.cache` as their cache folder,
and fix the ownership of any cache files that the test is directly
managing.

It also adds a `sudotest` target which runs the tests as root, and then
verifies that no root-owned files have been dropped into the cache.

Lastly, several setup() and cleanup() methods have been removed, as they
are largely unnecessary.  common-tap does all the required setup for
test package directories and cache folders, and removing the cache
avoids the root-ownership test, which is an important integration
requirement.