correct-mkdir: infer cache owner from parent dir
When running as sudo (without the -H flag), the HOME environment
variable is /var/root by default.

The correct-mkdir module was reading env.HOME to determine the
appropriate uid and gid for files written to the cache folder.

However, this is rarely correct.  In practice, this module was reading
env.SUDO_UID and env.SUDO_GID in some cases, but matching ownership on
/var/root in others.

This patch changes correct-mkdir to match the behavior of cacache v12,
where ownership of a file or folder is inferred from the ownership of
its nearest parent directory.