pacote@9.5.5
Infer the ownership of all unpacked files in node_modules, so that we
never have user-owned files in root-owned folders, or root-owned files
in user-owned folders.

This prevents one of the last remaining issues that bites users who use
`sudo` unnecessarily.  The only remaining issue is to no longer drop
perms when running scripts as root, but that is a breaking change which
will come in v7.