fix: Better error on invalid --user/--group
This addresses the issue when people fail to install binary packages on
Docker and other environments where there is no 'nobody' user.

The Config.loadUid() method was failing in these cases, but not in a way
that was particularly informative.  Furthermore, the uid and gid
resolved in that way were just ignored and never stored anywhere.

However, until npm-lifecycle@3.1.3, an error from uid-number was _also_
ignored, so if we didn't crash somewhere, then it would run scripts as
root when provided with an invalid user.  This is arguably fine, but it
is a violation of the contract that the npm CLI presents.

As of npm-lifecycle@3.1.3, these errors are handled properly in
npm-lifecycle, so the additional uninformative crash is no longer doing
anything.  This commit removes that uninformative crash.  This also
means that we won't fail _until_ an invalid user config is actually
relevant; if someone never runs an install script (or runs with
--ignore-scripts), then it's not relevant, so we can move forward
anyway.