config: save files with inferred ownership
Don't use SUDO_UID and SUDO_GID.  Just make the file match the folder
it's being written into.  The vast majority of the time, this will be
the user's home directory, but if it's not, then we should not leave a
user-owned file in a root-owned location.  And, if running as root
without SUDO_UID/SUDO_GID environs, but putting a config file in the
user's home dir, then it's quite rude to leave it root-owned.