Fix package id in shrinkwrap lifecycle step output
Currently all logging related to shrinkwrap steps reports 'undefined'
for the package in output and log messages.

This is due to the package associated with the `idealTree` being
recreated in the `savePackageJson()` method which precedes these
steps. For now, just copy forward the `_id` attribute which lifecycle
logging expects, but note that mutating `package` here is surprising.

Fixes npm/npm#20756

PR-URL: https://github.com/npm/cli/pull/288
Credit: @bz2
Close: #288
Reviewed-by: @claudiahdz