Tidy code for excluding readme from npm view
This worked but was very confusing because it was shadowing the
'args' variable with the version in the iterator function iterating
through itself, so the 'args' passed to `showFields` was the single
arg, not the array of args.

Moreover, it was performing the `indexOf` test on each individual
arg, so doing `indexOf` on the string rather than the array, which
still worked, but was somewhat confusing and probably not the
intention. Instead, check once outside of the loop to see if args
contains 'readme'.