Don't use tree.package._where when resolving shrinkwrap dependency
The path in `sw.requires` is always relative to the requesting package
location, and the default from `packageRelativePath(tree)` is always correct.

For `type=file` dependencies, `tree.package._where` contains the `realpath`
of the requesting tarball, which is the location where it was unpacked in
`node_modules`. That's completely useless for resolving dependency paths.
We're interested in the source location of the packed tarball. That's the path
against which the dependency spec should be resolved.