Do not open URL in a browser unless valid
This is a defense for all cases where we might be trying to open a web
url based on either a response from the server, or some form of user
input.  As far as we can tell, they're all being validated, but defense
in depth is always a good idea.