Do not ship CoC/Contributing files for deps
Some of these are used under a somewhat ambiguous license, and we're
moving away from the WeAll* stuff to just rely on the centralized npm
code of conduct anyhow.  Better to leave them out for now, as we go
through and update the deps themselves to have a cleaner and more
consistent project setup.