Better --force message, correct config.force documentation
The current warning is not an actual help message, is not suggesting any
actions, and it's quite condescending.

While the 'I sure hope you know what you are doing' was always intended
as a sincere and genuine hope, people tend to read it in a variety of
ways, and it doesn't provide any helpful information.

PR-URL: https://github.com/npm/cli/pull/1259
Credit: @pablocubico
Close: #1259
Close: #1244
Reviewed-by: @isaacs