Better --force message, correct config.force documentation
While the 'I sure hope you know what you are doing' was always intended
as a sincere and genuine hope, people tend to read it in a variety of
ways, and it doesn't provide much information.

This is an alternative to #1244, which I think is still a little judgey,
and not very informative.  Let's just tell people what's going on.

Close: #1244