[Breaking] add "exports" field
Since npm v7 is imminent, it's a great opportunity to add the "exports"
field, which is otherwise very hard to make nonbreaking.

I used the verbose form for ".", because node v13.0 and v13.1 require
the string fallback, and if you ever want to add an ESM wrapper, this
will make it trivial to add the "import" key to the object form
correctly. I used the "default" key because some node versions will
produce an unfortunate experimental warning if you use the "node" or
"require" keys.

I also included package.json since a ton of tools rely on this
information being requireable.

PR-URL: https://github.com/npm/cli/pull/1413
Credit: @ljharb
Close: #1413
Reviewed-by: @isaacs