docs: fix build error from libvips + sharp
Problem: Sharp doesn't play nicely with a globally-installed version of
libvips, and throws an error when you try to generate the documentation.
with `npm run build:static`. This has been documented as an issue in the
Gatsby issue tracker: https://github.com/gatsbyjs/gatsby/issues/22303

Solution: Update the dependencies to pull in a new version of Sharp
which is immune to this problem. I ran `npm update`, which is supposed
to modify `package.json`, but it only changed `package-lock.json`. I
removed `node_modules` and ran `npm install` and can confirm that his
resolves the problem.