Add 'npm exec', port npx to use it directly
This removes libnpx, and adds a new command, 'npm exec', which
implements the functionality.  As of this change going live, we are
dropping support for the standalone top-level package 'npx'.

Not all of the functionality of the old version of npx is maintained.

The shell fallback functionality is dropped.  It's insecure, and not
something we want to support or encourage.  If anyone wants it, they
can hack up their .bashrc file themselves.

--no-install is not supported.  If the package is not found locally, it
is installed in a predictable location in the cache, rather than
failing.  This is something we might want to review, as automatically
installing in the case of misspellings may be a security footgun.

--ignore-existing is dropped.  Existing packages are always given
priority.

--quiet or -q can be accomplished by using the --silent npm option, so
it's also dropped.

--npm option is dropped.  npx will always use the npm that it ships
with.

--node-arg is dropped.  There are other ways to set node options via
environment variables in the Node.js versions we support.

--always-spawn is dropped.  npx will always spawn a child process to
execute commands.

The --shell option can be accomplished by using the --script-shell npm
option.

--version and --help are just passed through to npm.

As an added bonus, I noticed that the files in `bin/` were not getting
run.  So now we have full coverage for npm-cli.js and npx-cli.js.

PR-URL: https://github.com/npm/cli/pull/1588
Credit: @isaacs
Close: #1588
Reviewed-by: @ruyadorno