Find any path to top in `isOnlyDev`
When determining whether a node is considered 'only' a development
dependency of the top-level, check whether there are any paths
from the specific node to the top-level through a direct
development dependency of the top-level to avoid issues where the
shortest path from the node to top-level is via a direct
production dependency, thus causing the node to be considered a
production dependency.

This comes into play particularly when issuing:

  npm install --only=development

as this will end up in a situation where transitive dependencies
of a development dependency may not be installed.