Add a 'check-coverage' script so we can track towards completion
We are targetting 100% test coverage for npm v7 GA.  Using a
coverage-map, we require that any of the newly created tests fully cover
the module that they are testing.  However, this has the side effect of
_always_ hitting 100% coverage when running 'npm test', even though not
all modules are being tested.

This adds a new 'load-all' test which, in 'check-coverage' mode, tells
nyc to include every file in the project.

This also removes the double-run of our tests in CI, where we run once
and then immediately run the same exact thing again for Coveralls, and
sends the 'check-coverage' output to Coveralls instead.