deps: float c2ccc1a on node-gyp
A change to "correctly rename object files for absolute paths"
caused a bug in windows. A fix has already landed upstream in
gyp-next, but has not yet made it's way to a release of gyp-next
or to node-gyp.

Let's float the change until it is fixed upstream.

Refs: https://github.com/nodejs/gyp-next/pull/74

PR-URL: https://github.com/npm/cli/pull/1974
Credit: @MylesBorins
Close: #1974
Reviewed-by: @ruyadorno