view: Better errors when package.json is not JSON
Previously, `npm view <noargs>` was just saying "Invalid package.json"
if the package.json file was not JSON, or if it was missing.

This errors out with the appropriate error in these cases.

Also, no need to read the big clunky 'read-package-json' for this, we're
literally just checking for a single field.  We can be a bit more
efficient here.