Support *all* conf keys in publishConfig
This adds a flatOptions.flatten() method, which takes an object full of
config keys, and turns it into an options object.  This method expects
an object that already inherits from npm's defaults, and is thus
expected to be internal only.

This commit also removes some config keys which were used by npm
dependencies at the start of the v7 beta process, but are no longer:

- all lockfile configs (since we don't use lockfiles any more! for
  anything! and good riddance, they're a rats' nest of race conditions)
- cacheMax/cacheMin (we only use preferOffline/offline/online now, so
  these are strictly legacy support as input and never shared with deps)

Once this lands in cli, we can remove the publishConfig handling logic
in npm-registry-fetch, as it will be redundant.