test: fix npm run test
Fixes two issues that enable running tests downstream in nodejs:
- Running tests using `npm run-script test` instead of the `npm test`
alias will result in a failure since `test/lib/npm.js` have an assertion
for "test" being the command only. This changes it so that this
particular test also accepts "run-script" as the command.
- Explicitly enable colors for tap output so that running the test suite
in a non-tty environment will result in output that matches coloured
snapshots.

PR-URL: https://github.com/npm/cli/pull/2225
Credit: @ruyadorno
Close: #2225
Reviewed-by: @darcyclarke