update all dependencies using latest npm to install them
This addresses the inclusion of some dev dependencies that should not
have been included, due to a bug in Arborist 1.x which has been
corrected.

It also just moves some stuff around that was still in the tree shape
that npm 6 created, which is different than what npm v7 will create.

Several formerly dev deps were moved to production dependencies due to
the addition of resolve in normalize-package-data v3.