ignore the sort in help-search more broadly
It turns out that the other stuff in those objects might also be in
random order, so this is still triggering CI coverage failures, albeit
more rarely than it used to.

Just ignore the whole sort function.  It's fine, we implicitly assert on
the sortedness in the test, so we know that it is doing its job.

PR-URL: https://github.com/npm/cli/pull/2574
Credit: @isaacs
Close: #2574
Reviewed-by: @nlf