chore(tests): consistent error assertions
Our handling of "there should be no error" was not consistent.  It is
preferable to stay within the assertion library of your testing
framework, but tap also does not by default bail out on a test suite on
a failed assertion.  The `bail` flag turns this on for these assertions,
so now we are dealing with them in a consistent way and in a way that
aligns with what people are likely expecting when they see this kind of
assertion (i.e. there should be no error).

Coincidentally this found a subtle bug in at least one test where we
were checking that no error was generated after we had called t.end().
The `if (err) throw err` not using the testing framework meant this was
impossible to catch.