fix: refactor get package name
renames read-local-package to read-package-name

The global check needed to be moved outside this function, because it
was handled differently (and will be even moreso when we implement diff
workspaces) in each function.  This allowed us to now pass in the prefix
itself instead of the npm object, so we can reuse this function to look
up package names when we refactor npm diff for workspaces.

PR-URL: https://github.com/npm/cli/pull/3331
Credit: @wraithgar
Close: #3331
Reviewed-by: @ruyadorno