fix(libnpmexec): don't detach output from npm
The npm output function refers to this.log.  libnpmexec has that passed
in, which decoupled the function from the npm object.  This fixes it,
and sets the tests up in a way where if the output function ever becomes
detached from this.npm in the same way, tests will fail.

PR-URL: https://github.com/npm/cli/pull/3329
Credit: @wraithgar
Close: #3329
Reviewed-by: @ruyadorno