chore(package) Use a "files" list
We have a fairly small number of things that we wish to ship in our
package, and they're all nicely arranged in a small number of folders.

Instead of having a sprawling .npmignore file, we should just get with
the recommended best practice, and use a "files" list in package.json.

PR-URL: https://github.com/npm/cli/pull/3365
Credit: @isaacs
Close: #3365
Reviewed-by: @wraithgar