fix(link): do not prune packages
`npm link <pkg>` is meant to be used as a way to link a local package to
an install tree and it's very surprising to users that it may prune
extraneous deps from the project.

This change switches the default behavior to avoid pruning deps when
reifying the dependencies in npm link.

Fixes: https://github.com/npm/cli/issues/2554

PR-URL: https://github.com/npm/cli/pull/3399
Credit: @ruyadorno
Close: #3399
Reviewed-by: @ljharb, @nlf