fix(config): add flatOptions.npxCache
This adds a new `npxCache` flatOption for libnpmexec to look for to put
its `_npx` content.  libnpmexec will have to be patched to use that
value, and continue to pass `flatOptions.cache` to pacote et al.

The `flatOptions.cache` is the one that is intended to be passed down
into other modules, as it has the `_cacache` suffix attached.  What was
happening before was that `npx` was creating a new alternate cache one
directory up from where everything else was, and also putting the `_npx`
content there.  It is possible this is the source of at least some of
our "npx doesn't find the right versions" bugs.

PR-URL: https://github.com/npm/cli/pull/3430
Credit: @wraithgar
Close: #3430
Reviewed-by: @ruyadorno