feat(ls): report *why* something is invalid
This is a papercut that has been driving me crazy when debugging
ERESOLVE issues.  It's not particularly useful to just say something is
"invalid", without showing which module's dependency is not being met.

With this commit, it prints all the packages that depend on that
dependency and do not have their required version met.

This does print multiple times for deduped deps that are invalid, but
if we skip the printing of the `invalid` label for deduped deps, we end
up losing information that is only detected later in the tree walk.

PR-URL: https://github.com/npm/cli/pull/3460
Credit: @isaacs
Close: #3460
Reviewed-by: @nlf