fix(tests): move more tests to use real npm
This moves a handful of the smaller tests to using the new npm mock that
uses the real actual npm object.  It also extends the testing surface
area of a few tests back down into the actual `process.spawn` that
results, instead of anything internal to the code.

Some dead code in `lib/test.js` was found during this, as well as an
instance of a module throwing a string instead of an error object.

PR-URL: https://github.com/npm/cli/pull/3463
Credit: @wraithgar
Close: #3463
Reviewed-by: @nlf