feat(workspaces): --include-workspace-root --no-workspaces
Adds a new config item that includes the workspace root when running
non-arborist commands (i.e. repo, version, publis).

This also changes --workspaces to a trinary, so that setting it to false
will explicitly exclude workspaces altogether.  This is also going to
require an arborist change so that it ignores workspaces altogether.